MODIFY: modify product/user models and add orders app
